+++ lib/graphlient/errors.rb
@@ -1,0 +1,2 @@
+require_relative 'errors/error'
+require_relative 'errors/http'